FIX: Main Tab, Habit View 수정
