+++ ContentView.swift
@@ -20,8 +20,8 @@ struct ContentView: View {
     var body: some View {
         MainView(
             store: Store(
-                initialState: MainFeature.State(),
-                reducer: { MainFeature() }
+                initialState: AppFeature.State(),
+                reducer: { AppFeature() }
             )
         )
         .environmentObject(setting)
